requirements.txt
# This file contains all development dependencies that are not provided
# file the Sovereign system.
marshmallow
coverage
nose
sousou.ioc
python-qpid-proton
alembic
